requirements.txt
pandas>=1.0.3
python-docx>=0.8.10
lxml>=4.5.1
natsort>=7.0.1